c/reg_test: make subprocess test libc-agnostic
A test in c/reg_test/run currently relies on glibc-specific
print buffering behavior in a test case that checks environment
vars with subprocesses.

The case in question made mlr spawn a script that prints a few lines
but also had mlr print lines itself at the same time. The order of
lines from mlr vs subprocess are undefined, producing a slightly
different outcome on Alpine Linux than on Glibc-based distributions.

This commit adds the "put -q" flag to the test case effectively
causing only the subprocess output to be printed (which is the one
the particular test case was interested in).

Relates to https://github.com/johnkerl/miller/issues/293