rename disposition-matrix shorthands to match Go port
In C I use left-justified formatting to make these line up for proofreading.
In Go there is gofmt which does not preserve that kind of formatting. A
workaround in Go is to rename the shorthand-functions to all have the same
number of characters.  While I'm porting from C to Go, it's to go ahead and
rename the C functions the same, so I can copy the disposition matrices from C
to Go with a minimum of reformatting or typo-making.