Compiling on newer go versions doesn't work
For whatever reason when compiling with go `1.23` it complains about
needing to `go tidy` and running that bumps the go version to `1.21` and
adds the `toolchain` directive, while also updating go.sum.

Compiling on go `1.20` works just fine without this update.

Not sure if you want to go all the way to `1.23` or do the minimum of
`1.21` so I just picked the latter and can change if you want to.